Before encodeArgs move.
