+++ internal/converter/payload_converter.go
@@ -17,3 +17,12 @@ type PayloadConverter interface {
 	// Encoding returns encoding supported by PayloadConverter.
 	Encoding() string
 }
+
+func newPayload(data []byte, c PayloadConverter) *commonpb.Payload {
+	return &commonpb.Payload{
+		Metadata: map[string][]byte{
+			metadataEncoding: []byte(c.Encoding()),
+		},
+		Data: data,
+	}
+}
